@@ -1,29 +1,35 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-	<PropertyGroup>
-		<TargetFramework>net5.0</TargetFramework>
-	</PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
 
-	<PropertyGroup>
-		<PackageId>SignalR.Orleans</PackageId>
-		<Title>SignalR.Orleans - Orleans backplane for SignalR Core</Title>
-		<Description>Orleans backplane for SignalR Core.</Description>
-		<PackageTags>$(PackageTags) AspNetCore</PackageTags>
-		<PackageReleaseNotes></PackageReleaseNotes>
-		<IncludeSymbols>true</IncludeSymbols>
-		<IncludeSource>true</IncludeSource>
-		<PackageVersion>2.0.1</PackageVersion>
-	</PropertyGroup>
+  <PropertyGroup>
+    <PackageId>SignalR.Orleans</PackageId>
+    <Title>SignalR.Orleans - Orleans backplane for SignalR Core</Title>
+    <Description>Orleans backplane for SignalR Core.</Description>
+    <PackageTags>$(PackageTags) AspNetCore</PackageTags>
+    <PackageReleaseNotes></PackageReleaseNotes>
+    <IncludeSymbols>true</IncludeSymbols>
+    <IncludeSource>true</IncludeSource>
+  </PropertyGroup>
 
-	<ItemGroup>
-		<!-- see https://gist.github.com/tebeco/a6e0d9a3885a0a36e702795219bd4fe9 -->
-		<FrameworkReference Include="Microsoft.AspNetCore.App" />
-		<PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" Version="$(OrleansCodeGeneratorVersion)">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-		<PackageReference Include="Microsoft.Orleans.Core" Version="$(OrleansCoreVersion)" />
-		<PackageReference Include="Microsoft.Orleans.OrleansProviders" Version="$(OrleansProvidersVersion)" />
-		<PackageReference Include="Microsoft.Orleans.Runtime.Abstractions" Version="$(OrleansRuntimeAbstractionsVersion)" />
-	</ItemGroup>
+  <ItemGroup>
+    <!-- see https://gist.github.com/tebeco/a6e0d9a3885a0a36e702795219bd4fe9 -->
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" Version="$(OrleansCodeGeneratorVersion)">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Orleans.Core" Version="$(OrleansCoreVersion)" />
+    <PackageReference Include="Microsoft.Orleans.OrleansProviders" Version="$(OrleansProvidersVersion)" />
+    <PackageReference Include="Microsoft.Orleans.Runtime.Abstractions" Version="$(OrleansRuntimeAbstractionsVersion)" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
+      <_Parameter1>SignalR.Orleans.Tests</_Parameter1>
+    </AssemblyAttribute>
+  </ItemGroup>
 </Project>
