@@ -1,42 +1,46 @@
-﻿using System.Buffers;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
+using System.Buffers;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.SignalR.Protocol;
+using System.Collections.Generic;
 using Microsoft.Extensions.Logging;
+using Microsoft.AspNetCore.SignalR.Protocol;
 using Orleans;
-using Orleans.Concurrency;
 using Orleans.Streams;
+using Orleans.Runtime;
+using Orleans.Concurrency;
 
 namespace SignalR.Orleans.Core
 {
-    internal abstract class ConnectionGrain<TGrainState> : Grain<TGrainState>, IConnectionGrain
+    internal abstract class ConnectionGrain<TGrainState> : Grain, IConnectionGrain
         where TGrainState : ConnectionState, new()
     {
         private readonly ILogger _logger;
-        private IStreamProvider _streamProvider;
-        private Dictionary<string, StreamSubscriptionHandle<string>> _connectionStreamHandles;
+        private readonly IPersistentState<TGrainState> _connectionState;
+        private readonly Dictionary<string, StreamSubscriptionHandle<string>> _connectionStreamHandles = new();
+        private IStreamProvider _streamProvider = default!;
 
         protected ConnectionGrainKey KeyData;
 
-        internal ConnectionGrain(ILogger logger)
+        internal ConnectionGrain(
+            ILogger logger,
+            IPersistentState<TGrainState> connectionState)
         {
             _logger = logger;
+            _connectionState = connectionState;
         }
 
         public override async Task OnActivateAsync()
         {
             KeyData = new ConnectionGrainKey(this.GetPrimaryKeyString());
-            _connectionStreamHandles = new Dictionary<string, StreamSubscriptionHandle<string>>();
             _streamProvider = GetStreamProvider(Constants.STREAM_PROVIDER);
             var subscriptionTasks = new List<Task>();
-            foreach (var connection in State.Connections)
+            foreach (var connection in _connectionState.State.Connections)
             {
                 var clientDisconnectStream = _streamProvider.GetStream<string>(Constants.CLIENT_DISCONNECT_STREAM_ID, connection);
                 var subscriptions = await clientDisconnectStream.GetAllSubscriptionHandles();
                 foreach (var subscription in subscriptions)
                 {
-                    subscriptionTasks.Add(subscription.ResumeAsync(async (connectionId, _) => await Remove(connectionId)));
+                    subscriptionTasks.Add(subscription.ResumeAsync((connectionId, _) => Remove(connectionId)));
                 }
             }
             await Task.WhenAll(subscriptionTasks);
@@ -44,53 +48,47 @@ namespace SignalR.Orleans.Core
 
         public virtual async Task Add(string connectionId)
         {
-            var shouldWriteState = State.Connections.Add(connectionId);
+            var shouldWriteState = _connectionState.State.Connections.Add(connectionId);
             if (!_connectionStreamHandles.ContainsKey(connectionId))
             {
                 var clientDisconnectStream = _streamProvider.GetStream<string>(Constants.CLIENT_DISCONNECT_STREAM_ID, connectionId);
-                var subscription = await clientDisconnectStream.SubscribeAsync(async (connId, _) => await Remove(connId));
+                var subscription = await clientDisconnectStream.SubscribeAsync((connId, _) => Remove(connId));
                 _connectionStreamHandles[connectionId] = subscription;
             }
 
             if (shouldWriteState)
-                await WriteStateAsync();
+                await _connectionState.WriteStateAsync();
         }
 
         public virtual async Task Remove(string connectionId)
         {
-            var shouldWriteState = State.Connections.Remove(connectionId);
+            var shouldWriteState = _connectionState.State.Connections.Remove(connectionId);
             if (_connectionStreamHandles.TryGetValue(connectionId, out var stream))
             {
                 await stream.UnsubscribeAsync();
                 _connectionStreamHandles.Remove(connectionId);
             }
 
-            if (State.Connections.Count == 0)
+            if (_connectionState.State.Connections.Count == 0)
             {
-                await ClearStateAsync();
+                await _connectionState.ClearStateAsync();
                 DeactivateOnIdle();
             }
             else if (shouldWriteState)
             {
-                await WriteStateAsync();
+                await _connectionState.WriteStateAsync();
             }
         }
 
-        public virtual Task Send(Immutable<InvocationMessage> message)
-        {
-            return SendAll(message, State.Connections);
-        }
+        public virtual Task Send(Immutable<InvocationMessage> message) => SendAll(message, _connectionState.State.Connections);
 
-        public Task SendExcept(string methodName, object[] args, IReadOnlyList<string> excludedConnectionIds)
+        public Task SendExcept(string methodName, object?[] args, IReadOnlyList<string> excludedConnectionIds)
         {
             var message = new Immutable<InvocationMessage>(new InvocationMessage(methodName, args));
-            return SendAll(message, State.Connections.Where(x => !excludedConnectionIds.Contains(x)).ToList());
+            return SendAll(message, _connectionState.State.Connections.Where(x => !excludedConnectionIds.Contains(x)).ToList());
         }
 
-        public Task<int> Count()
-        {
-            return Task.FromResult(State.Connections.Count);
-        }
+        public Task<int> Count() => Task.FromResult(_connectionState.State.Connections.Count);
 
         protected Task SendAll(Immutable<InvocationMessage> message, IReadOnlyCollection<string> connections)
         {
@@ -100,11 +98,11 @@ namespace SignalR.Orleans.Core
             var tasks = ArrayPool<Task>.Shared.Rent(connections.Count);
             try
             {
-                int index = 0;
-                foreach (var connection in connections)
+                for (int i = 0; i < connections.Count; i++)
                 {
+                    var connection = connections.ElementAt(i);
                     var client = GrainFactory.GetClientGrain(KeyData.HubName, connection);
-                    tasks[index++] = client.Send(message);
+                    tasks[i] = client.Send(message);
                 }
 
                 return Task.WhenAll(tasks.Where(x => x != null).ToArray());
@@ -118,6 +116,6 @@ namespace SignalR.Orleans.Core
 
     internal abstract class ConnectionState
     {
-        public HashSet<string> Connections { get; set; } = new HashSet<string>();
+        public HashSet<string> Connections { get; set; } = new();
     }
 }
