@@ -1,13 +1,12 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.SignalR.Protocol;
 using Microsoft.Extensions.Logging;
+using Microsoft.AspNetCore.SignalR.Protocol;
 using Orleans;
-using Orleans.Concurrency;
-using Orleans.Providers;
+using Orleans.Runtime;
 using Orleans.Streams;
+using Orleans.Concurrency;
 using SignalR.Orleans.Core;
 
 namespace SignalR.Orleans.Clients
@@ -20,23 +19,27 @@ namespace SignalR.Orleans.Clients
         public Guid ServerId { get; set; }
     }
 
-    [StorageProvider(ProviderName = Constants.STORAGE_PROVIDER)]
     [Reentrant]
-    internal class ClientGrain : Grain<ClientState>, IClientGrain
+    internal class ClientGrain : Grain, IClientGrain
     {
+        private const string CLIENT_STORAGE = "ClientState";
         private readonly ILogger<ClientGrain> _logger;
-        private IStreamProvider _streamProvider;
-        private IAsyncStream<ClientMessage> _serverStream;
-        private IAsyncStream<Guid> _serverDisconnectedStream;
-        private IAsyncStream<string> _clientDisconnectStream;
-        private ConnectionGrainKey _keyData;
-        private StreamSubscriptionHandle<Guid> _serverDisconnectedSubscription;
+        private readonly IPersistentState<ClientState> _clientState;
+        private IStreamProvider _streamProvider = default!;
+        private IAsyncStream<ClientMessage> _serverStream = default!;
+        private IAsyncStream<Guid> _serverDisconnectedStream = default!;
+        private IAsyncStream<string> _clientDisconnectStream = default!;
+        private ConnectionGrainKey _keyData = default!;
+        private StreamSubscriptionHandle<Guid>? _serverDisconnectedSubscription;
         private const int _maxFailAttempts = 3;
         private int _failAttempts;
 
-        public ClientGrain(ILogger<ClientGrain> logger)
+        public ClientGrain(
+            ILogger<ClientGrain> logger,
+            [PersistentState(CLIENT_STORAGE, Constants.STORAGE_PROVIDER)] IPersistentState<ClientState> clientState)
         {
             _logger = logger;
+            _clientState = clientState;
         }
 
         public override async Task OnActivateAsync()
@@ -45,27 +48,29 @@ namespace SignalR.Orleans.Clients
             _streamProvider = GetStreamProvider(Constants.STREAM_PROVIDER);
             _clientDisconnectStream = _streamProvider.GetStream<string>(Constants.CLIENT_DISCONNECT_STREAM_ID, _keyData.Id);
 
-            if (State.ServerId == Guid.Empty)
+            if (_clientState.State.ServerId == Guid.Empty)
                 return;
 
-            _serverStream = _streamProvider.GetStream<ClientMessage>(State.ServerId, Constants.SERVERS_STREAM);
-            _serverDisconnectedStream = _streamProvider.GetStream<Guid>(State.ServerId, Constants.SERVER_DISCONNECTED);
+            _serverStream = _streamProvider.GetStream<ClientMessage>(_clientState.State.ServerId, Constants.SERVERS_STREAM);
+            _serverDisconnectedStream = _streamProvider.GetStream<Guid>(_clientState.State.ServerId, Constants.SERVER_DISCONNECTED);
             var subscriptions = await _serverDisconnectedStream.GetAllSubscriptionHandles();
-            var subscriptionTasks = new List<Task>();
-            foreach (var subscription in subscriptions)
+            var subscriptionTasks = new Task[subscriptions.Count];
+            for (int i = 0; i < subscriptions.Count; i++)
             {
-                subscriptionTasks.Add(subscription.ResumeAsync(async (serverId, _) => await OnDisconnect("server-disconnected")));
+                var subscription = subscriptions[i];
+                subscriptionTasks[i] = subscription.ResumeAsync((serverId, _) => OnDisconnect("server-disconnected"));
             }
+
             await Task.WhenAll(subscriptionTasks);
         }
 
         public async Task Send(Immutable<InvocationMessage> message)
         {
-            if (State.ServerId != Guid.Empty)
+            if (_clientState.State.ServerId != Guid.Empty)
             {
                 _logger.LogDebug("Sending message on {hubName}.{targetMethod} to connection {connectionId}", _keyData.HubName, message.Value.Target, _keyData.Id);
                 _failAttempts = 0;
-                await _serverStream.OnNextAsync(new ClientMessage { ConnectionId = _keyData.Id, Payload = message.Value, HubName = _keyData.HubName });
+                await _serverStream.OnNextAsync(new ClientMessage(_keyData.HubName, _keyData.Id, message.Value));
                 return;
             }
 
@@ -82,23 +87,23 @@ namespace SignalR.Orleans.Clients
 
         public async Task OnConnect(Guid serverId)
         {
-            State.ServerId = serverId;
-            _serverStream = _streamProvider.GetStream<ClientMessage>(State.ServerId, Constants.SERVERS_STREAM);
-            _serverDisconnectedStream = _streamProvider.GetStream<Guid>(State.ServerId, Constants.SERVER_DISCONNECTED);
-            _serverDisconnectedSubscription = await _serverDisconnectedStream.SubscribeAsync(async _ => await OnDisconnect("server-disconnected"));
-            await WriteStateAsync();
+            _clientState.State.ServerId = serverId;
+            _serverStream = _streamProvider.GetStream<ClientMessage>(_clientState.State.ServerId, Constants.SERVERS_STREAM);
+            _serverDisconnectedStream = _streamProvider.GetStream<Guid>(_clientState.State.ServerId, Constants.SERVER_DISCONNECTED);
+            _serverDisconnectedSubscription = await _serverDisconnectedStream.SubscribeAsync(_ => OnDisconnect("server-disconnected"));
+            await _clientState.WriteStateAsync();
         }
 
-        public async Task OnDisconnect(string reason = null)
+        public async Task OnDisconnect(string? reason = null)
         {
             _logger.LogDebug("Disconnecting connection on {hubName} for connection {connectionId} from server {serverId} via {reason}",
-                _keyData.HubName, _keyData.Id, State.ServerId, reason);
+                _keyData.HubName, _keyData.Id, _clientState.State.ServerId, reason);
 
             if (_keyData.Id != null)
             {
                 await _clientDisconnectStream.OnNextAsync(_keyData.Id);
             }
-            await ClearStateAsync();
+            await _clientState.ClearStateAsync();
 
             if (_serverDisconnectedSubscription != null)
                 await _serverDisconnectedSubscription.UnsubscribeAsync();
